Increased functionality
Increase the overall website functionality by adding the ability to purchase a product. This gives the template for all future products. The top-up and purchase queries were also changed to search the database for the user's id instead of a previously static id. This means that any user should now be able to register, login, topup and purchase products.